+++ lib/internal/modules/esm/loader.js
@@ -31,6 +31,7 @@ const {
   ERR_UNKNOWN_MODULE_FORMAT
 } = require('internal/errors').codes;
 const { pathToFileURL, isURLInstance, URL } = require('internal/url');
+const { emitExperimentalWarning } = require('internal/util');
 const {
   isAnyArrayBuffer,
   isArrayBufferView,
@@ -106,6 +107,18 @@ class ESMLoader {
    * Methods which translate input code or other information into ES modules
    */
   translators = translators;
+
+  constructor() {
+    if (getOptionValue('--experimental-loader')) {
+      emitExperimentalWarning('Custom ESM Loaders');
+    }
+    if (getOptionValue('--experimental-network-imports')) {
+      emitExperimentalWarning('Network Imports');
+    }
+    if (getOptionValue('--experimental-specifier-resolution') === 'node') {
+      emitExperimentalWarning('Specifier Resolution');
+    }
+  }
 
   static pluckHooks({
     globalPreload,
