+++ lib/internal/modules/esm/resolve.js
@@ -364,7 +364,6 @@ function resolveDirectoryEntry(search) {
 }
 
 const encodedSepRegEx = /%2F|%5C/i;
-let experimentalSpecifierResolutionWarned = false;
 /**
  * @param {URL} resolved
  * @param {string | URL | undefined} base
@@ -379,13 +378,6 @@ function finalizeResolution(resolved, base, preserveSymlinks) {
 
   let path = fileURLToPath(resolved);
   if (getOptionValue('--experimental-specifier-resolution') === 'node') {
-    if (!experimentalSpecifierResolutionWarned) {
-      process.emitWarning(
-        'The Node.js specifier resolution flag is experimental. It could change or be removed at any time.',
-        'ExperimentalWarning');
-      experimentalSpecifierResolutionWarned = true;
-    }
-
     let file = resolveExtensionsWithTryExactName(resolved);
 
     // Directory
