+++ lib/internal/process/esm_loader.js
@@ -55,9 +55,6 @@ async function initializeLoader() {
 
   if (!customLoaders.length) return;
 
-  const { emitExperimentalWarning } = require('internal/util');
-  emitExperimentalWarning('--experimental-loader');
-
   let cwd;
   try {
     cwd = process.cwd() + '/';
